fix(scripting): move backend-scripting toast to its own module to break an import cycle
Importing showBackendScriptingDisabledToast / BackendScriptingDisabledError as
values from frontend_script_api pulled its widget graph into entrypoints and
bundle at init time, tripping a circular-import error (note_context_aware_widget
extends an undefined BasicWidget) that failed ~18 client test suites. Move both
into a lightweight services/backend_scripting.js module that all three consumers
import instead, so the widget graph is no longer eagerly loaded.
