+++ apps/client/src/services/bundle.ts
@@ -4,7 +4,8 @@ import { h, VNode } from "preact";
 import FNote from "../entities/fnote.js";
 import BasicWidget, { ReactWrappedWidget } from "../widgets/basic_widget.js";
 import RightPanelWidget from "../widgets/right_panel_widget.js";
-import { BackendScriptingDisabledError, type Entity } from "./frontend_script_api.js";
+import { BackendScriptingDisabledError } from "./backend_scripting.js";
+import type { Entity } from "./frontend_script_api.js";
 import { WidgetDefinitionWithType } from "./frontend_script_api_preact.js";
 import { t } from "./i18n.js";
 import ScriptContext from "./script_context.js";
