+++ apps/client/src/services/frontend_script_api.spec.ts
@@ -3,9 +3,10 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import appContext from "../components/app_context.js";
 import type FNote from "../entities/fnote.js";
 import { buildNote } from "../test/easy-froca.js";
+import { BackendScriptingDisabledError } from "./backend_scripting.js";
 import dateNotesService from "./date_notes.js";
 import dialogService from "./dialog.js";
-import FrontendScriptApi, { BackendScriptingDisabledError, type Api, type Entity } from "./frontend_script_api.js";
+import FrontendScriptApi, { type Api, type Entity } from "./frontend_script_api.js";
 import { preactAPI } from "./frontend_script_api_preact.js";
 import froca from "./froca.js";
 import linkService from "./link.js";
